Added syncproducts, mapper and db caching

+++ src/main/java/org/example/model/Product.java
@@ -5,7 +5,7 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 @Entity
 @Getter
@@ -20,7 +20,7 @@ public class Product {
     private String title;
 
     @Column
-    private BigInteger price;
+    private BigDecimal price;
 
     @Column
     private String description;
